+++ src/extension.ts
@@ -3,8 +3,9 @@ import { openMermaidPreview } from "./commands/openMermaidPreview";
 import { openMermaidSource } from "./commands/openMermaidSource";
 import { relativeGoTo } from "./commands/relativeGoTo";
 import { activateReviewCommentController } from "./git/reviewCommentController";
-import { exportReviewSummary, promptAndAddReviewNote, publishReviewToPR } from "./git/publishReview";
+import { exportReviewSummary, promptAndAddReviewNote } from "./git/publishReview";
 import { getActiveRepository } from "./git/resolveActiveRepository";
+import { ActionPanelViewProvider } from "./actionPanel/ActionPanelView";
 import { GitHelpersViewProvider } from "./gitHelpers/GitHelpersView";
 import { registerMermaidCustomEditor } from "./mermaidEditor/MermaidCustomEditorProvider";
 import { SnippetViewProvider } from "./snippetEditor/SnippetView";
@@ -28,19 +29,13 @@ export function activate(context: ExtensionContext) {
     registerMermaidCustomEditor(context);
     SnippetViewProvider.activate(context);
     activateReviewCommentController(context);
+    ActionPanelViewProvider.activate(context);
     GitHelpersViewProvider.activate(context);
 
     registerCommandIB(Commands.AddReviewNote, async () => {
         const repo = await getActiveRepository();
         if (repo) {
             await promptAndAddReviewNote(repo.rootUri.fsPath);
-        }
-    }, context);
-    registerCommandIB(Commands.PublishReviewToPR, async () => {
-        const repo = await getActiveRepository();
-        const branch = repo?.state.HEAD?.name;
-        if (repo && branch) {
-            await publishReviewToPR(repo.rootUri.fsPath, branch);
         }
     }, context);
     registerCommandIB(Commands.ExportReviewSummary, async () => {